# BE/api-spec.yml
openapi: "3.0.2"
info:
  title: Anti Retake API
  version: '1.2'
servers:
  - url: https://antiretake.xyz/api
paths:
...
  /sets/{studySetId}/save:
    post:
      summary: Save study set to folder.
      tags:
        - Study Set
      operationId: API-031
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: studySetId
          required: true
          schema:
            type: string
        - in: query
          name: folderId
          schema:
            type: string
      responses:
        "201":
          description: Save study set to folder successful.
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ar_Auth